+++ packages/blockchain-link/src/workers/blockbook/handlers/getAccountBalanceHistory.ts
@@ -1,0 +1,17 @@
+import type { MessageTypes, ResponseTypes as Responses } from '@trezor/blockchain-link-types';
+import { RESPONSES } from '@trezor/blockchain-link-types';
+
+import type { Request } from '../types';
+
+export const getAccountBalanceHistory = async (
+    request: Request<MessageTypes.GetAccountBalanceHistory>,
+): Promise<Responses.GetAccountBalanceHistory> => {
+    const { payload } = request;
+    const api = await request.connect();
+    const history = await api.getAccountBalanceHistory(payload);
+
+    return {
+        type: RESPONSES.GET_ACCOUNT_BALANCE_HISTORY,
+        payload: history,
+    };
+};
