+++ packages/blockchain-link/src/workers/blockbook/index.ts
@@ -1,425 +1,26 @@
-import {
-    type BlockbookAddressNotification as AddressNotification,
-    type BlockbookBlockNotification as BlockNotification,
-    CustomError,
-    type BlockbookFiatRatesNotification as FiatRatesNotification,
-    MESSAGES,
-    type BlockbookMempoolTransactionNotification as MempoolTransactionNotification,
-    type MessageTypes,
-    RESPONSES,
-    type Response,
-    type SubscriptionAccountInfo,
-} from '@trezor/blockchain-link-types';
-import * as utils from '@trezor/blockchain-link-utils/src/blockbook';
+import { CustomError, MESSAGES, RESPONSES } from '@trezor/blockchain-link-types';
+import type { MessageTypes, Response } from '@trezor/blockchain-link-types';
 
-import { BaseWorker, CONTEXT, type ContextType } from '../baseWorker';
+import { BaseWorker, CONTEXT } from '../baseWorker';
+import { estimateFee } from './handlers/estimateFee';
+import { getAccountBalanceHistory } from './handlers/getAccountBalanceHistory';
+import { getAccountInfo } from './handlers/getAccountInfo';
+import { getAccountUtxo } from './handlers/getAccountUtxo';
+import { getBlock } from './handlers/getBlock';
+import { getBlockHash } from './handlers/getBlockHash';
+import { getContractInfo } from './handlers/getContractInfo';
+import { getCurrentFiatRates } from './handlers/getCurrentFiatRates';
+import { getFiatRatesForTimestamps } from './handlers/getFiatRatesForTimestamps';
+import { getFiatRatesTickersList } from './handlers/getFiatRatesTickersList';
+import { getInfo } from './handlers/getInfo';
+import { getTransaction } from './handlers/getTransaction';
+import { getTransactionHex } from './handlers/getTransactionHex';
+import { pushTransaction } from './handlers/pushTransaction';
+import { rpcCall } from './handlers/rpcCall';
+import { subscribe } from './handlers/subscribe';
+import { unsubscribe } from './handlers/unsubscribe';
+import type { Request } from './types';
 import { BlockbookAPI } from './websocket';
-
-type Context = ContextType<BlockbookAPI>;
-type Request<T> = T & Context;
-
-const getInfo = async (request: Request<MessageTypes.GetInfo>) => {
-    const api = await request.connect();
-    const info = await api.getServerInfo();
-
-    return {
-        type: RESPONSES.GET_INFO,
-        payload: {
-            url: api.options.url,
-            ...utils.transformServerInfo(info),
-        },
-    } as const;
-};
-
-const getBlockHash = async (request: Request<MessageTypes.GetBlockHash>) => {
-    const api = await request.connect();
-    const info = await api.getBlockHash(request.payload);
-
-    return {
-        type: RESPONSES.GET_BLOCK_HASH,
-        payload: info.hash,
-    } as const;
-};
-
-const getBlock = async (request: Request<MessageTypes.GetBlock>) => {
-    const api = await request.connect();
-    const info = await api.getBlock(request.payload);
-
-    return {
-        type: RESPONSES.GET_BLOCK,
-        payload: info,
-    } as const;
-};
-
-const getAccountInfo = async (request: Request<MessageTypes.GetAccountInfo>) => {
-    const { payload } = request;
-    const api = await request.connect();
-    const info = await api.getAccountInfo(payload);
-
-    return {
-        type: RESPONSES.GET_ACCOUNT_INFO,
-        payload: utils.transformAccountInfo(info),
-    } as const;
-};
-
-const getAccountUtxo = async (request: Request<MessageTypes.GetAccountUtxo>) => {
-    const { payload } = request;
-    const api = await request.connect();
-    const utxos = await api.getAccountUtxo(payload);
-
-    return {
-        type: RESPONSES.GET_ACCOUNT_UTXO,
-        payload: utils.transformAccountUtxo(utxos),
-    } as const;
-};
-
-const getAccountBalanceHistory = async (
-    request: Request<MessageTypes.GetAccountBalanceHistory>,
-) => {
-    const { payload } = request;
-    const api = await request.connect();
-    const history = await api.getAccountBalanceHistory(payload);
-
-    return {
-        type: RESPONSES.GET_ACCOUNT_BALANCE_HISTORY,
-        payload: history,
-    } as const;
-};
-
-const getCurrentFiatRates = async (request: Request<MessageTypes.GetCurrentFiatRates>) => {
-    const { payload } = request;
-    const api = await request.connect();
-    const fiatRates = await api.getCurrentFiatRates(payload);
-
-    return {
-        type: RESPONSES.GET_CURRENT_FIAT_RATES,
-        payload: fiatRates,
-    } as const;
-};
-
-const getFiatRatesForTimestamps = async (
-    request: Request<MessageTypes.GetFiatRatesForTimestamps>,
-) => {
-    const { payload } = request;
-    const api = await request.connect();
-    const { tickers } = await api.getFiatRatesForTimestamps(payload);
-
-    return {
-        type: RESPONSES.GET_FIAT_RATES_FOR_TIMESTAMPS,
-        payload: { tickers },
-    } as const;
-};
-
-const getFiatRatesTickersList = async (request: Request<MessageTypes.GetFiatRatesTickersList>) => {
-    const { payload } = request;
-    const api = await request.connect();
-    const tickers = await api.getFiatRatesTickersList(payload);
-
-    return {
-        type: RESPONSES.GET_FIAT_RATES_TICKERS_LIST,
-        payload: {
-            ts: tickers.ts,
-            availableCurrencies: tickers.available_currencies, // convert to camelCase
-        },
-    } as const;
-};
-
-const getTransaction = async (request: Request<MessageTypes.GetTransaction>) => {
-    const api = await request.connect();
-    const { txid, descriptor } = request.payload;
-    const rawtx = await api.getTransaction(txid);
-    const account = descriptor ? request.state.getAccount(descriptor) : undefined;
-    const tx = utils.transformTransaction(
-        rawtx,
-        account?.addresses ?? account?.descriptor ?? descriptor,
-    );
-
-    return {
-        type: RESPONSES.GET_TRANSACTION,
-        payload: tx,
-    } as const;
-};
-
-const getTransactionHex = async (request: Request<MessageTypes.GetTransactionHex>) => {
-    const api = await request.connect();
-    const { hex } = await api.getTransaction(request.payload);
-    if (!hex) throw new CustomError(`Missing hex of ${request.payload}`);
-
-    return {
-        type: RESPONSES.GET_TRANSACTION_HEX,
-        payload: hex,
-    } as const;
-};
-
-const pushTransaction = async (request: Request<MessageTypes.PushTransaction>) => {
-    const api = await request.connect();
-    const { hex, disableAlternativeRPC } = request.payload;
-    const resp = await api.pushTransaction(hex, disableAlternativeRPC);
-
-    return {
-        type: RESPONSES.PUSH_TRANSACTION,
-        payload: resp.result,
-    } as const;
-};
-
-const estimateFee = async (request: Request<MessageTypes.EstimateFee>) => {
-    const api = await request.connect();
-    const resp = await api.estimateFee(request.payload);
-
-    return {
-        type: RESPONSES.ESTIMATE_FEE,
-        payload: resp,
-    } as const;
-};
-
-const rpcCall = async (request: Request<MessageTypes.RpcCall>) => {
-    const api = await request.connect();
-    const resp = await api.rpcCall(request.payload);
-
-    return {
-        type: RESPONSES.RPC_CALL,
-        payload: resp,
-    } as const;
-};
-
-const getContractInfo = async (request: Request<MessageTypes.GetContractInfo>) => {
-    const api = await request.connect();
-    const response = await api.getContractInfo(request.payload);
-
-    return {
-        type: RESPONSES.GET_CONTRACT_INFO,
-        payload: response,
-    } as const;
-};
-
-const onNewBlock = ({ post }: Context, event: BlockNotification) => {
-    post({
-        id: -1,
-        type: RESPONSES.NOTIFICATION,
-        payload: {
-            type: 'block',
-            payload: {
-                blockHeight: event.height,
-                blockHash: event.hash,
-                evmData: event.evmData ?? null,
-            },
-        },
-    });
-};
-
-const onMempoolTx = ({ post }: Context, payload: MempoolTransactionNotification) => {
-    post({
-        id: -1,
-        type: RESPONSES.NOTIFICATION,
-        payload: {
-            type: 'mempool',
-            payload,
-        },
-    });
-};
-
-const onTransaction = ({ state, post }: Context, event: AddressNotification) => {
-    if (!event.tx) return;
-    const descriptor = event.address;
-    // check if there is subscribed account with received address
-    const account = state.getAccount(descriptor);
-    post({
-        id: -1,
-        type: RESPONSES.NOTIFICATION,
-        payload: {
-            type: 'notification',
-            payload: {
-                descriptor: account ? account.descriptor : descriptor,
-                tx: account
-                    ? utils.transformTransaction(event.tx, account.addresses ?? account.descriptor)
-                    : utils.transformTransaction(event.tx, descriptor),
-            },
-        },
-    });
-};
-
-const onNewFiatRates = ({ post }: Context, event: FiatRatesNotification) => {
-    post({
-        id: -1,
-        type: RESPONSES.NOTIFICATION,
-        payload: {
-            type: 'fiatRates',
-            payload: {
-                rates: event.rates,
-            },
-        },
-    });
-};
-
-const subscribeAccounts = async (ctx: Context, accounts: SubscriptionAccountInfo[]) => {
-    // subscribe to new blocks, confirmed and mempool transactions for given addresses
-    const api = await ctx.connect();
-    const { state } = ctx;
-    state.addAccounts(accounts);
-    if (!state.getSubscription('notification')) {
-        api.on('notification', ev => onTransaction(ctx, ev));
-        state.addSubscription('notification');
-    }
-
-    return api.subscribeAddresses(state.getAddresses());
-};
-
-const subscribeAddresses = async (ctx: Context, addresses: string[]) => {
-    // subscribe to new blocks, confirmed and mempool transactions for given addresses
-    const api = await ctx.connect();
-    const { state } = ctx;
-    state.addAddresses(addresses);
-    if (!state.getSubscription('notification')) {
-        api.on('notification', ev => onTransaction(ctx, ev));
-        state.addSubscription('notification');
-    }
-
-    return api.subscribeAddresses(state.getAddresses());
-};
-
-const subscribeBlock = async (ctx: Context) => {
-    if (ctx.state.getSubscription('block')) return { subscribed: true };
-    ctx.state.addSubscription('block');
-    const api = await ctx.connect();
-    api.on('block', ev => onNewBlock(ctx, ev));
-
-    return api.subscribeBlock();
-};
-
-const subscribeFiatRates = async (ctx: Context, currency?: string) => {
-    const api = await ctx.connect();
-    if (!ctx.state.getSubscription('fiatRates')) {
-        ctx.state.addSubscription('fiatRates');
-        api.on('fiatRates', ev => onNewFiatRates(ctx, ev));
-    }
-
-    return api.subscribeFiatRates(currency);
-};
-
-const subscribeMempool = async (ctx: Context) => {
-    const api = await ctx.connect();
-    if (!ctx.state.getSubscription('mempool')) {
-        ctx.state.addSubscription('mempool');
-        api.on('mempool', ev => onMempoolTx(ctx, ev));
-    }
-
-    return api.subscribeMempool();
-};
-
-const subscribe = async (request: Request<MessageTypes.Subscribe>) => {
-    const { payload } = request;
-
-    let response: { subscribed: boolean };
-    if (payload.type === 'accounts') {
-        response = await subscribeAccounts(request, payload.accounts);
-    } else if (payload.type === 'addresses') {
-        response = await subscribeAddresses(request, payload.addresses);
-    } else if (payload.type === 'block') {
-        response = await subscribeBlock(request);
-    } else if (payload.type === 'fiatRates') {
-        response = await subscribeFiatRates(request, payload.currency);
-    } else if (payload.type === 'mempool') {
-        response = await subscribeMempool(request);
-    } else {
-        throw new CustomError('invalid_param', '+type');
-    }
-
-    return {
-        type: RESPONSES.SUBSCRIBE,
-        payload: response,
-    } as const;
-};
-
-const unsubscribeAccounts = async (
-    { state, connect }: Context,
-    accounts?: SubscriptionAccountInfo[],
-) => {
-    state.removeAccounts(accounts || state.getAccounts());
-
-    const api = await connect();
-    const subscribed = state.getAddresses();
-    if (subscribed.length < 1) {
-        // there are no subscribed addresses left
-        // remove listeners
-        api.removeAllListeners('notification');
-        state.removeSubscription('notification');
-
-        return api.unsubscribeAddresses();
-    }
-
-    // subscribe remained addresses
-    return api.subscribeAddresses(subscribed);
-};
-
-const unsubscribeAddresses = async ({ state, connect }: Context, addresses?: string[]) => {
-    const api = await connect();
-    // remove accounts
-    if (!addresses) {
-        state.removeAccounts(state.getAccounts());
-    }
-    const subscribed = state.removeAddresses(addresses || state.getAddresses());
-    if (subscribed.length < 1) {
-        // there are no subscribed addresses left
-        // remove listeners
-        api.removeAllListeners('notification');
-        state.removeSubscription('notification');
-
-        return api.unsubscribeAddresses();
-    }
-
-    // subscribe remained addresses
-    return api.subscribeAddresses(subscribed);
-};
-
-const unsubscribeBlock = async ({ state, connect }: Context) => {
-    if (!state.getSubscription('block')) return { subscribed: false };
-    const api = await connect();
-    api.removeAllListeners('block');
-    state.removeSubscription('block');
-
-    return api.unsubscribeBlock();
-};
-
-const unsubscribeFiatRates = async ({ state, connect }: Context) => {
-    if (!state.getSubscription('fiatRates')) return { subscribed: false };
-    const api = await connect();
-    api.removeAllListeners('fiatRates');
-    state.removeSubscription('fiatRates');
-
-    return api.unsubscribeFiatRates();
-};
-
-const unsubscribeMempool = async ({ state, connect }: Context) => {
-    if (!state.getSubscription('mempool')) return { subscribed: false };
-    const api = await connect();
-    api.removeAllListeners('mempool');
-    state.removeSubscription('mempool');
-
-    return api.unsubscribeMempool();
-};
-
-const unsubscribe = async (request: Request<MessageTypes.Unsubscribe>) => {
-    const { payload } = request;
-    let response: { subscribed: boolean };
-    if (payload.type === 'accounts') {
-        response = await unsubscribeAccounts(request, payload.accounts);
-    } else if (payload.type === 'addresses') {
-        response = await unsubscribeAddresses(request, payload.addresses);
-    } else if (payload.type === 'block') {
-        response = await unsubscribeBlock(request);
-    } else if (payload.type === 'fiatRates') {
-        response = await unsubscribeFiatRates(request);
-    } else if (payload.type === 'mempool') {
-        response = await unsubscribeMempool(request);
-    } else {
-        throw new CustomError('invalid_param', '+type');
-    }
-
-    return {
-        type: RESPONSES.UNSUBSCRIBE,
-        payload: response,
-    } as const;
-};
 
 const onRequest = (request: Request<MessageTypes.Message>) => {
     switch (request.type) {
