+++ packages/blockchain-link/src/workers/blockbook/types.ts
@@ -1,0 +1,6 @@
+import type { ContextType } from '../baseWorker';
+import type { BlockbookAPI } from './websocket';
+
+export type Context = ContextType<BlockbookAPI>;
+
+export type Request<T> = T & Context;
